fixed all memory leaks

+++ namelib.h
@@ -14,7 +14,7 @@ typedef void* Element;
 typedef char* Name;
 typedef int Data;
 typedef const char* CName;
-typedef float Score;
+typedef double Score;
